requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
pyodbc==5.2.0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
